Build sgxsdk derivation in sandbox
When using nix-build the environment is "sandboxed" and as a consequence
/usr/bin/env is not available and scripts having a shebang with
/usr/bin/env must be patched. Basically everything resides under the nix
store (/nix/store/...) and therefore any reference to commands or
utilities with a path that is not under the nix store will result in an
error. For instance invoking /bin/cp will fail as the cp used in the
sandbox resides under the nix store (/nix/store/...).

The the nixpkgs/nix docker image uses some tricks to get around the
above sandbox characteristics such that using nix-build may work in the
image whereas it won't on the host machine.